feat: mock data type 업데이트 및 shared type 파일 생성
