+++ api/types/review.ts
@@ -1,18 +1,4 @@
-import { TCategory } from "./menu";
-
-export interface IImage {
-  id: number;
-  url: string;
-}
-
-export type TSource = "외부사이트" | "당도";
-
-export type TReviewOption =
-  | "맛있어요"
-  | "응답이 빨라요"
-  | "가성비가 좋아요"
-  | "친절해요"
-  | "선물하기 좋아요";
+import { IImage, TCategory, TReviewOption, TSource } from "./shared";
 
 export interface IReviewItem {
   nickname: string;
